Abo roqaya tasks for journals in backend

+++ BLL/Models/Tracking/Jornal/EditJournal.cs
@@ -10,5 +10,5 @@ public class EditJournal
     [StringLength(10000, ErrorMessage = "Content cannot exceed 10000 characters")]
     public string Content { get; set; } = string.Empty;
 
-    public decimal Mood { get; set; }
+    public float Mood { get; set; }
 }
